FIX: bug in dimmlevels because of int size

+++ src/lightcontrol.cpp
@@ -1,8 +1,8 @@
 #include "lightcontrol.h"
 
 // ZERO terminated list of available dimmLevels
-int dimmLevels[] = {8, 16, 32, 64, 128, 255};
-int dimmLevel = 8;
+byte dimmLevels[] = {8, 16, 32, 64, 128, 255};
+byte dimmLevel = 8;
 
 bool lightIsOn = false;
 
@@ -54,7 +54,7 @@ void lightDimm() {
   MySerial.println(dimmLevel);
   MySerial.print(F("levels: "));
   MySerial.println(sizeof(dimmLevels));
-  
+
   if(dimmLevel == dimmLevels[sizeof(dimmLevels) - 1]) {
     // Current value is the last one
     // Delay a little
